store: Remove JSONB filters
